fix(globe): make the Predict button work in 3D
The button did nothing. It called window.openSatellitePredict, which is
registered by the Leaflet satellite layer — the layer WorldMap skips
entirely in 3D, so the global never existed there and the guard swallowed
every click. My earlier claim that Predict "came across intact because it
is a global" was wrong: the global only exists when that plugin mounts.

Extracted the prediction modal to utils/satellitePredict.js, taking its
dependencies (satellites, config, t) as arguments instead of closing over
plugin state. The plugin now delegates to it and still registers the
global for existing callers, so the Flat path is unchanged; the globe
calls it directly and no longer cares whether any plugin has run.

That also removes a hidden coupling: a button in one component depending
on a side effect of a plugin mounting elsewhere.

Verified both paths: in 3D the modal opens with real pass times while
window.openSatellitePredict is still undefined, and in Flat the plugin
window's Predict opens the same modal with the global registered as
before.
